refactor(dmr): split client.go into focused files by concern
Split the monolithic pkg/model/provider/dmr/client.go (1,236 lines) into
five focused files organized by concern:

- client.go: Core Client struct, NewClient, CreateChatCompletionStream
- resolve.go: URL resolution, fallback logic, connectivity testing
- configure.go: Model configuration HTTP call, runtime flag merging
- embed.go: CreateEmbedding, CreateBatchEmbedding, Rerank
- pull.go: Model pulling, existence check, user confirmation

Also simplify code across the split files: extract URL-construction
helpers, delegate CreateEmbedding to CreateBatchEmbedding, streamline
flag parsing, and remove duplicate assignments (~216 net line reduction).

Fix e2e tests failing when DOCKER_CLI_PLUGIN_ORIGINAL_CLI_COMMAND is
inherited from the host environment by clearing it in runCLI.
